add: fps counter

+++ include/utils.hpp
@@ -33,4 +33,4 @@ void setPixel(SDL_Renderer *renderer ,int x, int y, SDL_Color c);
 
 void Circle_draw(SDL_Renderer *renderer, Circle *circle);
 
-int getFPS();
+float getFPS();
